New Tool - AutoLogoff. replaces the logoff20minutesafterlock thing

+++ logoff20minlock/OnWorkstationUnlockTask.ps1
@@ -1,3 +1,5 @@
+if ($env:USERNAME -eq "Administrator") { exit }
+
 $OWLpid = [Environment]::GetEnvironmentVariable("OWLpid", "User")
 
 Stop-Process -Id $OWLpid -Force
